requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.2
django-coverage-plugin==1.3.1
Sphinx==1.4.5
django-extensions==1.6.7
Werkzeug==0.11.10
django-test-plus==1.0.14
factory_boy==2.7.0
django-debug-toolbar==1.5

# improved REPL
ipdb==0.10.1

pytest-django==2.9.1
pytest-sugar==0.7.1

# Your custom requirements go here

# Requirements for cookiecutter-django-crud
django-vanilla-views==1.0.4
django-bootstrap3==7.0.1
model-mommy==1.2.6
webtest==2.0.23
django-webtest==1.7.9

# For form input creation
django-localflavor==1.3

# WYSIWYG editor for model fields
django-wysiwyg-redactor==0.4.9